commandHauls: extract appending haul cols, to module
